# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Concert Deal Finder API",
    "description": "API for finding the best concert deals combining ticket prices with travel and accommodation costs",
    "version": "1.0.0",
    "contact": {
      "name": "Concert Deal Finder"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.concertdeals.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/tours": {
      "get": {
        "summary": "Search for concert tour dates",
        "description": "Find upcoming concert dates for a given artist",
        "parameters": [
          {
            "name": "artist",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "description": "Artist name to search for"
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "description": "City for origin location"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tour dates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Show"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Show": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "lat": {
            "type": "number"
          },
          "lon": {
            "type": "number"
          },
          "minPrice": {
            "type": "number"
          },
          "maxPrice": {
            "type": "number"
          },
          "ticketUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["VALIDATION_ERROR"]
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}